requirements.txt
pyserial==2.7
docopt==0.6.1
